tests: remove bash script, use Perl instead
The shell invocation only works (on Ubuntu 15.10) if $(SHELL) is
set to /bin/bash; otherwise the string replacement fails.

Instead of adding a dependency on bash, use Perl, since we already
know we have Perl in the system.

While at it, remove the dead link to prev/foo.png images, and
the spurious </body> tags.

Signed-off-by: Emilio G. Cota <cota@braap.org>